+++ Git2SemVer.MSBuild/Versioning/Generation/Builders/DefaultVersionBuilder.cs
@@ -4,7 +4,7 @@ using NoeticTools.Git2SemVer.MSBuild.Framework.BuildHosting;
 using NoeticTools.Git2SemVer.MSBuild.Framework.Config;
 using NoeticTools.Git2SemVer.MSBuild.Framework.Semver;
 using NoeticTools.Git2SemVer.MSBuild.Scripting;
-using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistory;
+using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistoryWalking;
 using Semver;
