+++ Git2SemVer.MSBuild/Versioning/Generation/Builders/DefaultVersionBuilderFactory.cs
@@ -1,5 +1,5 @@
 ﻿using NoeticTools.Common.Logging;
-using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistory;
+using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistoryWalking;
 
 
 namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.Builders;
