+++ Git2SemVer.MSBuild/Versioning/Generation/GitHistoryWalking/VersionHistoryPathExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Immutable;
 
 
-namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistory;
+namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistoryWalking;
 
 #pragma warning disable CS1591
