+++ Git2SemVer.MSBuild/Versioning/Generation/GitHistoryWalking/VersionHistoryPathsBuilder.cs
@@ -1,8 +1,7 @@
-﻿using System.Text;
-using NoeticTools.Common.Logging;
+﻿using NoeticTools.Common.Logging;
 
 
-namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistory;
+namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistoryWalking;
 
 #pragma warning disable CS1591
 internal sealed class VersionHistoryPathsBuilder
