+++ Git2SemVer.MSBuild/Versioning/Generation/GitHistoryWalking/VersionHistorySegment.cs
@@ -1,14 +1,12 @@
-﻿using Microsoft.Build.Utilities;
-using NoeticTools.Common.Exceptions;
+﻿using NoeticTools.Common.Exceptions;
 using NoeticTools.Common.Logging;
 using NoeticTools.Common.Tools.Git;
 using Semver;
-using Spectre.Console.Rendering;
 
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistory;
+namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistoryWalking;
 
 internal sealed class VersionHistorySegment
 {
