+++ Git2SemVer.MSBuild/Versioning/Generation/GitHistoryWalking/VersionHistorySegmentsBuilder.cs
@@ -1,10 +1,9 @@
 ﻿using System.Diagnostics;
-using System.Xml.Schema;
 using NoeticTools.Common.Logging;
 using NoeticTools.Common.Tools.Git;
 
 
-namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistory;
+namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistoryWalking;
 
 internal sealed class VersionHistorySegmentsBuilder
 {
