+++ Git2SemVer.MSBuild/Versioning/Generation/IGitHistoryPathsFinder.cs
@@ -1,4 +1,4 @@
-﻿using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistory;
+﻿using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistoryWalking;
 
 
 namespace NoeticTools.Git2SemVer.MSBuild.Versioning.Generation;
