+++ Git2SemVer.MSBuild/Versioning/GitOutputs.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.Json.Serialization;
 using NoeticTools.Common.Tools.Git;
 using NoeticTools.Git2SemVer.MSBuild.Framework.BuildHosting;
-using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistory;
+using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.GitHistoryWalking;
 using Semver;
